+++ client/public/index.html
@@ -1,5 +1,5 @@
 <!doctype html>
-<html lang="en">
+<html lang="en-us">
 
 <head>
   <meta charset="utf-8">
