+++ client/src/App.js
@@ -1,11 +1,13 @@
 import React from "react";
-import Signup from "./components/sign-up.js"
+import Signup from "./components/Signup.js";
+import Video from "./components/Video.js";
 
 function App() {
   return (
     <div>
       <h1>Welcome to GrowthTracker</h1>
       <Signup />
+      <Video />
     </div>
   );
 }
