+++ package.json
@@ -23,7 +23,8 @@
     "axios": "^0.18.1",
     "express": "^4.16.3",
     "if-env": "^1.0.4",
-    "mongoose": "^5.3.16"
+    "mongoose": "^5.3.16",
+    "prop-types": "^15.7.2"
   },
   "repository": {
     "type": "git",
